% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultisitePVA.R
\name{calculate_params_from_file}
\alias{calculate_params_from_file}
\title{Extract population parameters from a data file}
\usage{
calculate_params_from_file(data_file)
}
\arguments{
\item{data_file}{The path to the file}
}
\value{
A \code{list} containing initial_pops, growth_rate_corrs,
growth_rate_means and growth_rate_vars parameters. Please see \link{simulate_ms_pva} for a description of each parameter
}
\description{
This function expects a n_pop+1 columns layout with column headers as the first line.
First column should be years, then counts for each population. E.g. :\cr\cr
Year;Pop1;Pop2\cr
2010;35;26\cr
2011;26;41\cr
2012;33;35
}
\details{
This function calculates log(lambda) means and variances the standard way,
unless there are gaps in the time series.
In the latter case, regression analysis is used.

See Morris & Doak 2002, p.64-69 for calculation details
}
\examples{
## @knitr paramsExample

params <- calculate_params_from_file(
 system.file("extdata", "PolarBear_Stirling2004.csv", package = "msPVA")
)
res <- do.call(
 "simulate_ms_pva",
 c(
   params,
   list(
     K = c(300,200),
     leaving_prob = 0.1,
     reaching_prob = 0.7,
     quasi_extinction_thresholds = c(20, 20),
     n_years = 50,
     n_runs = 100,
     with_progress_bar = FALSE
   )
 )
)
print(res)
}

