% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultisitePVA.R
\name{simulate_ss_pva}
\alias{simulate_ss_pva}
\title{Run a stochastic simulation for a count-based single-site population viability analysis}
\usage{
simulate_ss_pva(initial_pops, n_years = 100, n_runs = 1000, K = NA,
  quasi_extinction_thresholds = 0, ...)
}
\arguments{
\item{initial_pops}{Number. Initial population sizes.}

\item{n_years}{A number. How many years should we simulate?}

\item{n_runs}{A number. How many simulations should we do?}

\item{K}{A number. Maximum population size (carrying capacity).}

\item{quasi_extinction_thresholds}{A number. Near extinction threshold for the population.}

\item{...}{Either lambdas (a vector of yearly growth rates), log_lambdas (a vector of log(lambdas) calculated as in Morris & Doak 2002, p.64-65) or growth_rate_means and growth_rate_vars (Numbers: Mean and variance of log(lambdas))}
}
\value{
A list-based S3 object of class \code{ssPVARes} containing elements final_pops (vector), n_years, n_runs, initial_pops, decline_risk and extinction_risk.
}
\description{
This function uses either a stochastic algorithm if log-lambdas or mean log-lambdas and variance
are provided, or a bootstrap method if lambdas are provided directly.
}
\details{
This function is not yet vectorized, so provided a single population per run.
}
\examples{
## @knitr ssExample

res <- simulate_ss_pva(
 growth_rate_means = 0.043,
 growth_rate_vars = 0.051,
 initial_pops = 70,
 K = 286,
 quasi_extinction_thresholds = 20,
 n_years = 50,
 n_runs = 100
)

print(res)
hist(res)
}

